% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVsplitEval.R
\name{CVsplitEval}
\alias{CVsplitEval}
\title{Evaluate the performance of binary classifiers}
\usage{
CVsplitEval(splitvar, X, targets, data)
}
\arguments{
\item{splitvar}{A character string specifying the split criteria (e.g. "age < 60").}

\item{X}{A dataframe with predictor variables.}

\item{targets}{A data frame containing target variables.}

\item{data}{A data frame containing the predictor and response variables.}
}
\value{
A list with the performance metrics for each target variable
}
\description{
A function that evaluates the performance of binary classifiers based on data split criteria
}
\details{
The function takes four inputs - \code{splitvar}, \code{X}, \code{targets}
and \code{data} - and outputs a list with measures of the performance of binary
classifiers. It first splits the data into two groups based on the specified split
criteria, and then fits binary classifiers to the resulting data. The performance
of the classifiers is then evaluated using various measures, such as accuracy and
kappa, and is stored in the output list.

The function first splits the data into two groups based on the split criteria
specified in \code{splitvar}. The binary classifiers are then fit to the resulting data,
and their performance is evaluated using various measures such as accuracy, kappa, and
root mean squared error. The results are stored in a list and returned as the output
of the function.

Note that the function uses the \code{glm} function in R to fit the binary classifiers, and
the accuracy, kappa, and root mean squared error are calculated using the \code{accuracy},
\code{kappa2}, and \code{rmse} functions, respectively.
}
\examples{
data(mutate_example)
features <- c("age", "sex", "biomarker")
outcomes <- c("response", "tumor_size", "ae_count", "OS_definition_time_status")
outcome_defs <- c("Cat", "Cont", "Count", "Surv")
targets <- list(Definitions = outcome_defs, Z = mutate_example[, outcomes])

result <- CVsplitEval("age < 60", mutate_example[, features], targets, mutate_example)
}
\seealso{
Other relevant R functions include \code{glm}, \code{accuracy}, \code{kappa2}, and \code{rmse}.
}
