% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPartSummary.R
\name{MTPartSummary}
\alias{MTPartSummary}
\title{MTPartSummary Function}
\usage{
MTPartSummary(tree)
}
\arguments{
\item{tree}{A list representing a decision tree generated by a tree-based
algorithm in R.}
}
\value{
A matrix containing summary statistics including the number of splits,
number of leaves, complexity parameter (CP), average relative error,
total relative error, cross-validated error (Xerror), standard
deviation of cross-validated error (Xstd), and evaluation metric (eval).
}
\description{
Extracts summary statistics from a decision tree.
}
\examples{
data(mutate_example)
features <- c("age", "sex", "biomarker")
outcomes <- c("response", "tumor_size", "ae_count", "OS_definition_time_status")
outcome_defs <- c("Cat", "Cont", "Count", "Surv")

tree <- MTPart(features, outcomes, outcome_defs, mutate_example[1:150, ],
               depth = 2, nodesize = 30)
summ <- MTPartSummary(tree)
summ$summary_table
}
