% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPrune.R
\name{MTPrune}
\alias{MTPrune}
\title{MTPrune: Prune a Decision Tree Using Minimal Cost-Complexity Pruning}
\usage{
MTPrune(tree, cp = 0.02)
}
\arguments{
\item{tree}{A decision tree object generated by a tree-building algorithm, such as \code{rpart} or \code{ctree}.}

\item{cp}{A numeric value specifying the complexity parameter (CP) threshold for pruning. Default is 0.02.}
}
\value{
A list with two elements:
\itemize{
\item \code{partitions}: A data frame representing the pruned parent-child relationships and tree nodes.
\item \code{tree_nodes}: A list of pruned decision tree objects, one for each node in the original tree.
}
}
\description{
Prune a MTPart decision tree object using minimal cost-complexity pruning based on the specified complexity parameter (CP) value.
}
\examples{
data(mutate_example)
features <- c("age", "sex", "biomarker")
outcomes <- c("response", "tumor_size", "ae_count", "OS_definition_time_status")
outcome_defs <- c("Cat", "Cont", "Count", "Surv")

tree <- MTPart(features, outcomes, outcome_defs, mutate_example[1:150, ],
               depth = 2, nodesize = 30)
pruned <- MTPrune(tree, cp = 0.02)
}
