% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTSummary.R
\name{MTSummary}
\alias{MTSummary}
\title{Summarize multiple outcome variables}
\usage{
MTSummary(targets, data)
}
\arguments{
\item{targets}{A data frame of targets with two columns, Definitions and Z. Definitions are defined as character
vectors indicating the type of variable: "Cont" for continuous, "Cat" for categorical, "Surv" for survival, and "Count" for count data.
Z is a matrix of outcomes with each row representing a patient and each column representing a target.}

\item{data}{A data frame of patient-level data with columns matching the names in targets.}
}
\value{
A list of lists with each element corresponding to an outcome in targets.
}
\description{
This function takes in a data frame of outcomes and a target table with definitions
for each target, and outputs summary statistics for each target. It currently
supports continuous, categorical, survival, and count outcome variable types.
}
\details{
The \code{MTSummary()} function computes a summary of the outcomes in a dataset for multiple
types of variables, including categorical, survival and count variables. The targets
argument is a list containing a vector of outcome definitions and a matrix or data
frame of outcomes. The data argument is a matrix or data frame with the data to be summarized.

For continuous targets, the function computes the mean, median, inter-quartile range, and MSE.

For categorical targets, the function computes the counts and proportions of each
category, the most frequent category and its proportion, and the expected loss (i.e.,
proportion of observations that do not belong to the most frequent category).

For survival targets, the function computes the number of missing values, the number
of events, the estimated rate of the event, and the deviance.

For count targets, the function computes the number of missing values, the total count,
the estimated rate of the count, and the deviance.
}
\examples{
data(mutate_example)
outcomes <- c("response", "tumor_size", "ae_count", "OS_definition_time_status")
outcome_defs <- c("Cat", "Cont", "Count", "Surv")
targets <- list(Definitions = outcome_defs, Z = mutate_example[, outcomes])
summ <- MTSummary(targets, mutate_example)
}
