% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTTest.R
\name{MTTest}
\alias{MTTest}
\title{MTTest: Test Multi-Target Tree}
\usage{
MTTest(tree, features, outcomes, outcome_defs, data_test, wt = NULL)
}
\arguments{
\item{tree}{A list containing the trained tree model.}

\item{features}{A character vector specifying the names of the features used in the tree model.}

\item{outcomes}{A character vector specifying the names of the outcomes used in the tree model.}

\item{outcome_defs}{A list containing definitions of the outcomes used in the tree model.}

\item{data_test}{A data frame containing the test data.}

\item{wt}{An optional numeric vector of weights for the test data.}
}
\value{
A list of evaluation results for the test data.
}
\description{
A function to perform tree-based evaluation on test data using a trained tree model.
}
\examples{
data(mutate_example)
features <- c("age", "sex", "biomarker")
outcomes <- c("response", "tumor_size", "ae_count", "OS_definition_time_status")
outcome_defs <- c("Cat", "Cont", "Count", "Surv")

train <- mutate_example[1:150, ]
tree <- MTPart(features, outcomes, outcome_defs, train,
               depth = 2, nodesize = 30)
test_result <- MTTest(tree, features, outcomes, outcome_defs, train)
}
