% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVTune.R
\name{CV_Tune}
\alias{CV_Tune}
\title{CV_Tune: Cross-Validated Hyperparameter Tuning for MTPart Models}
\usage{
CV_Tune(features, outcomes, outcome_defs, data,
        continuous, quantseq, wt, reuse, kfolds, Y, seedno,
        drange, noderange, splitmin_div, method, alpharange,
        igrange, psplitrange, pdepthrange, cp_val)
}
\arguments{
\item{features}{A character vector of feature names.}

\item{outcomes}{A character vector of outcome names.}

\item{outcome_defs}{A list of outcome definitions.}

\item{data}{A data frame containing the input data.}

\item{continuous}{A string indicating the method for handling continuous outcomes. Default is "quantile".}

\item{quantseq}{A numeric vector specifying the quantile sequence for quantile-based methods. Default is seq(0,1,0.25).}

\item{wt}{A numeric vector of weights for the samples. Default is NULL.}

\item{reuse}{A logical value indicating whether to reuse previous MTPart model fits. Default is FALSE.}

\item{kfolds}{An integer specifying the number of folds for k-fold cross-validation. Default is 10.}

\item{Y}{A character string specifying the outcome variable for partitioning.}

\item{seedno}{An integer specifying the seed for random number generation. Default is 1234.}

\item{drange}{A numeric vector specifying the range of values for the 'depth' parameter. Default is seq(2, 6, by=2).}

\item{noderange}{A numeric vector specifying the range of values for the 'nodesize' parameter. Default is seq(10, 30, by=10).}

\item{splitmin_div}{A numeric vector specifying the range of values for the 'splitmin_div' parameter. Default is seq(2, 4, by=1).}

\item{method}{A character vector specifying the method for selecting the best split. Default is c("avgIG","maxIG","mostIG","avgPVal","minPVal","mostPVal","splitError").}

\item{alpharange}{A numeric vector specifying the range of values for the 'alpha' parameter. Default is seq(0.05, 0.15, by=0.05).}

\item{igrange}{A numeric vector specifying the range of values for the 'IGcutoff' parameter. Default is seq(0.85, 0.95, by=0.05).}

\item{psplitrange}{A numeric vector specifying the range of values for the 'psplit' parameter. Default is seq(1, 2, by=1).}

\item{pdepthrange}{A numeric vector specifying the range of values for the 'pdepth' parameter. Default is seq(1, 2, by=1).}

\item{cp_val}{A numeric vector specifying the range of values for the 'CP' parameter. Default is c(seq(-1, 0, by=0.5)).}
}
\value{
A data frame containing the results of hyperparameter tuning for the MTPart model.
}
\description{
This function performs hyperparameter tuning for MTPart models using k-fold cross-validation.
}
