% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPart.R
\name{MTPart}
\alias{MTPart}
\title{Construct multi-target decision trees}
\usage{
MTPart(
  features,
  outcomes,
  outcome_defs,
  data,
  continuous = "quantile",
  quantseq = seq(0, 1, 0.25),
  wt = NULL,
  evalmethod = "avgIG",
  alpha = 0.05,
  IGcutoff = 0.95,
  depth = 4,
  nodesize = 20,
  cp = -1,
  reuse = FALSE,
  parallelpart = FALSE,
  parallelsplit = NA,
  paralleldepth = NA,
  splitmin = floor(nodesize/2)
)
}
\arguments{
\item{features}{Vector containing feature names to be evaluated.}

\item{outcomes}{Data frame object of all outcomes being evaluated. Used only in \code{Cat_Eval,
Cont_Eval, Count_Eval, Surv_Eval}.}

\item{outcome_defs}{Vector containing a list of outcome types ("Cat", "Cont", "Count", "Surv").}

\item{data}{The full dataset to be evaluated.}

\item{continuous}{How to handle partition evaluation of continuous features, in which
all potential split candidates may be examined ("all") or split candidates are derived
from quantiles ("quantile").}

\item{quantseq}{A user specified sequence ranging from 0 to 1 from which quantiles may
be specified (ex: seq(0,1,0.05)). If \code{continuous} and \code{quantseq} are not
specified the function will default to the "all" setting for continuous variables with
fewer than 100 unique values. If there are more than 100 unique values, the function
will use the "quantile" option with the default setting of \code{seq(0,1,0.05)}.}

\item{wt}{A numeric vector of the outcome variable weights to be applied in multi-target evaluations.
Default weights are set to \code{NULL}, indicating equal target variable importance.}

\item{evalmethod}{The method used to evaluate standardized proportion of information
gain across targets and target types.}

\item{alpha}{A numeric scalar specifying the significance level to be used in \code{evalmethod} if a p-value based
method (such as ) is selected. The default value for \code{alpha} is set to \code{0.05}.}

\item{IGcutoff}{A numeric scalar specifying the IG cutoff threshold to be used for \code{evalmethod} if methods
xxx or yyyy are selected. The default value for \code{IGcutoff} is \code{0.95}.}

\item{depth}{An integer scalar specifying the maximum depth of the tree (default is \code{6})}

\item{nodesize}{An integer scalar specifying the minimum number of samples per node (default is \code{20})}

\item{cp}{A numeric scalar specifying the complexity parameter for pruning (default is \code{0.02})}

\item{reuse}{A logical scalar specifying whether to reuse split variables at higher levels of the tree (default is \code{FALSE})}

\item{parallelpart}{A logical scalar specifying whether to use parallel computing for the partitioning step (default is \code{TRUE})}

\item{parallelsplit}{An integer scalar specifying the number of cores to use for parallel computing (default is \code{2})}

\item{paralleldepth}{An integer scalar specifying the minimum depth at which to start using parallel computing (default is \code{2})}

\item{splitmin}{An integer scalar specifying the minimum sample size \code{n} required for a node to be partitioned.
This is passed internally from \code{MTPart} and is used to assess feasibility of
feature binarization for evaluation (i.e., a binarized feature must result in
partitions with sufficient observations in both child nodes). The default value is of
\code{splitmin} is set to half of the \code{nodesize}.}
}
\description{
Evaluates multiple outcome variables of different types for a set of features.
}
