% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTPartSummary.R
\name{MTPartSummary}
\alias{MTPartSummary}
\title{MTPartSummary Function}
\usage{
MTPartSummary(tree)
}
\arguments{
\item{tree}{A list representing a decision tree generated by a tree-based
algorithm in R.}
}
\value{
A matrix containing summary statistics including the number of splits,
number of leaves, complexity parameter (CP), average relative error,
total relative error, cross-validated error (Xerror), standard
deviation of cross-validated error (Xstd), and evaluation metric (eval).
}
\description{
Extracts summary statistics from a decision tree.
}
