% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{PlotTree}
\alias{PlotTree}
\title{PlotTree: Plot Decision Tree}
\usage{
PlotTree(tree)
}
\arguments{
\item{tree}{A decision tree object.}
}
\value{
A plot of the multi-target decision tree.
}
\description{
Plot a multi-target decision tree.
}
