% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitPrep.R
\name{SplitPrep}
\alias{SplitPrep}
\title{Prepare data for splitting in a decision tree}
\usage{
SplitPrep(Xdf, df, data_splt, parentsplit)
}
\arguments{
\item{Xdf}{A data frame containing the predictor variables.}

\item{df}{A data frame containing the response variable.}

\item{data_splt}{The data used for the split.}

\item{parentsplit}{The parent split.}
}
\value{
A list containing two data frames: the modified predictor variable
data frame and the modified response variable data frame.
}
\description{
\code{SplitPrep()} is a function that prepares data frames for use in tree-based
models by identifying and removing candidate predictor variables that are unsuitable for
splitting at a given node. This function takes in three data frames, \code{Xdf}, \code{df},
and \code{data_splt} and a character string \code{parentsplit}.
It returns a list of two data frames: \code{splitX_df}, which contains the predictor variables
that can be used for splitting at a given node, and \code{split_data_df}, which contains the response
variables and any other variables that should be included in the data frame for further analysis.
}
\details{
\code{SplitPrep()} takes in two data frames, one containing the predictor variables
and another containing the response variable. It identifies which predictor
variables to use for splitting in the decision tree and returns a list
containing two data frames: the modified predictor variable data frame and the
modified response variable data frame.

The function first drops any unused factor levels in both input data frames
(\code{Xdf} and \code{df}) using the \code{droplevels()} function. It then loops through the predictor
variables in \code{Xdf} to determine which ones should be kept for splitting at the current node.

For each predictor variable, the function first checks if it was used for splitting at the
parent node (identified by the parentsplit argument). If so, it is added to a list of
variables to drop (\code{droplist}).

If the predictor variable is categorical (i.e., a factor variable), the function checks if
it has less than 2 levels or a prevalence of less than 5\% in the current node. If either of
these conditions is true, the variable is added to \code{droplist}.

If the predictor variable is continuous, the function checks if it has only one observed
value. If so, it is added to \code{droplist}.

Finally, if \code{droplist} exists, the function removes the corresponding columns from \code{splitX_df}
and renames the remaining columns to match the original column names. The function then
returns a list containing \code{splitX_df} and \code{split_data_df}.

This function is used to pre-process data before fitting tree-based models to ensure
that only suitable predictor variables are used for splitting.
}
